requirements.txt
black==24.10.0
click==8.1.8
beautifulsoup4==4.12.3
bs4==0.0.2
exceptiongroup==1.2.2
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pydub==0.25.1
pytest==8.3.4
python-docx==1.1.2
python-pptx==1.0.2
soupsieve==2.6
tomli==2.2.1
typing_extensions==4.12.2
XlsxWriter==3.2.2
